app2/app/requirements.txt
flask
psycopg2-binary
redis
python-dotenv
